+++ src/generators/index.ts
@@ -2,8 +2,9 @@
  * Generator module exports
  */
 
-export * from './base';
-export * from './object-generator';
-export * from './enum-generator';
-export * from './primitive-generator';
-export * from './factory';
+export * from "./base";
+export * from "./object-generator";
+export * from "./enum-generator";
+export * from "./primitive-generator";
+export * from "./factory";
+export * from "./hook-generator";